feat: switch to deadpool-sqlite

+++ reso/src/api/pagination.rs
@@ -2,15 +2,15 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize)]
 pub struct PagedQuery {
-    skip: Option<usize>,
-    top: Option<usize>,
+    skip: Option<i64>,
+    top: Option<i64>,
 }
 
 impl PagedQuery {
-    pub fn skip(&self) -> usize {
+    pub fn skip(&self) -> i64 {
         self.skip.unwrap_or(0)
     }
-    pub fn top(&self) -> usize {
+    pub fn top(&self) -> i64 {
         self.top.unwrap_or(25)
     }
 }
@@ -18,16 +18,16 @@ impl PagedQuery {
 #[derive(Serialize, Debug)]
 pub struct PagedResponse<T: Serialize> {
     pub items: Vec<T>,
-    pub total: usize,
-    pub top: usize,
-    pub skip: usize,
+    pub total: i64,
+    pub top: i64,
+    pub skip: i64,
     pub has_more: bool,
-    pub next_offset: usize,
+    pub next_offset: i64,
 }
 
 impl<T: Serialize> PagedResponse<T> {
-    pub fn new(items: Vec<T>, total: usize, top: usize, skip: usize) -> Self {
-        let next_offset = skip.saturating_add(items.len());
+    pub fn new(items: Vec<T>, total: i64, top: i64, skip: i64) -> Self {
+        let next_offset = skip.saturating_add(items.len() as i64);
         let has_more = next_offset < total;
 
         Self {
